Only release JDBC connection if entity manager was pre-bound
In case of pre-bound entity manager it makes sense to release connection
after transaction (in some cases at least).

But if entity manager is not pre-bound, then it is going to be closed
immediately after transaction cleanup. And that releases the connection.
So there's no point in releasing the connection manually, if it's
going to be released next moment anyway.