Allow Hibernate users release connections after transaction in case of open-in-view
This commit introduces a new setting - releaseConnectionAfterTransaction, which
controls if HibernateJpaDialect#releaseJdbcConnection will release connections
when called. ReleaseJdbcConnection method is only called if session is pre-bound,
so the new setting actually doesn't do anything unless open-in-view is on.

And in case if open-in-view is on, turning releaseConnectionAfterTransaction on
might help with connection pool starvation.